# Conflicts:
#   package.json

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsPvcServiceImpl.java
@@ -426,7 +426,7 @@ public class DevopsPvcServiceImpl implements DevopsPvcService {
         int level = ResourceUnitLevelEnum.valueOf(unit.toUpperCase()).ordinal();
 
         // 1024的一次方 对应ki 1024的2次方 对应Mi 以此类推
-        size = (long) (size * Math.pow(1024, level + 2));
+        size = (long) (size * Math.pow((double) 1024, (double) level + 2));
 
         BigDecimal bigDecimal = new BigDecimal(size);
         Quantity quantity = new Quantity(bigDecimal, Quantity.Format.BINARY_SI);
